requirements.txt
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
aws-cdk.assets==1.13.1
aws-cdk.aws-apigateway==1.11.0
aws-cdk.aws-applicationautoscaling==1.11.0
aws-cdk.aws-autoscaling==1.11.0
aws-cdk.aws-autoscaling-common==1.11.0
aws-cdk.aws-autoscaling-hooktargets==1.11.0
aws-cdk.aws-certificatemanager==1.11.0
aws-cdk.aws-cloudformation==1.11.0
aws-cdk.aws-cloudfront==1.11.0
aws-cdk.aws-cloudwatch==1.13.1
aws-cdk.aws-codebuild==1.11.0
aws-cdk.aws-codecommit==1.11.0
aws-cdk.aws-codepipeline==1.11.0
aws-cdk.aws-ec2==1.13.1
aws-cdk.aws-ecr==1.11.0
aws-cdk.aws-ecr-assets==1.11.0
aws-cdk.aws-ecs==1.11.0
aws-cdk.aws-ecs-patterns==1.11.0
aws-cdk.aws-elasticloadbalancing==1.11.0
aws-cdk.aws-elasticloadbalancingv2==1.11.0
aws-cdk.aws-events==1.13.1
aws-cdk.aws-events-targets==1.11.0
aws-cdk.aws-iam==1.13.1
aws-cdk.aws-kms==1.13.1
aws-cdk.aws-lambda==1.13.1
aws-cdk.aws-logs==1.13.1
aws-cdk.aws-route53==1.11.0
aws-cdk.aws-route53-targets==1.11.0
aws-cdk.aws-s3==1.13.1
aws-cdk.aws-s3-assets==1.13.1
aws-cdk.aws-s3-notifications==1.13.1
aws-cdk.aws-secretsmanager==1.11.0
aws-cdk.aws-servicediscovery==1.11.0
aws-cdk.aws-sns==1.13.1
aws-cdk.aws-sns-subscriptions==1.11.0
aws-cdk.aws-sqs==1.13.1
aws-cdk.aws-ssm==1.13.1
aws-cdk.aws-stepfunctions==1.11.0
aws-cdk.core==1.13.1
aws-cdk.cx-api==1.13.1
aws-cdk.region-info==1.13.1
boto3==1.9.252
botocore==1.12.252
cattrs==0.9.0
docutils==0.15.2
-e git+git@github.com:Flood-ctrl/cdk-ec2-instance.git@df89b70a9898ba1c31b41ded493b4596d72cee9e#egg=ec2_instance
importlib-metadata==0.23
isort==4.3.21
jmespath==0.9.4
jsii==0.19.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
publication==0.0.3
py==1.8.0
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
